* All:
Add `artwork` to genres. (✔️|✔️|✔️)
Add `sortType`, `orderType` and `ignoreCase` to `queryAudiosFrom`. (✔️|✔️|✔️)

* Dart:
Add `frameBuilder` and `errorBuilder` to `[QueryArtworkWidget].` (✔️)
Remove deprecated methods. (✔️)
Update DEPRECATED. (✔️)

* Web
Add `path` parameter to `querySongs`. (✔️)

* Android:
Downgrade kotlin and gradle version. (✔️)
Fix null album_id on Android 9 or below. (✔️)
Re-Add the `path` parameter to `querySongs`. (✔️)
Fix empty `Uint8List` when using `queryArtwork` on Android 7. (✔️)

+++ on_audio_query/ios/Classes/query/helper/OnAudioHelper.swift
@@ -176,7 +176,7 @@ public func formatGenreList(args: [String: Any], allGenres: [[String: Any?]]) ->
     return tempList
 }
 
-public func getMediaCount(type: Int, id: Int) -> Int {
+public func getMediaCount(type: Int, id: UInt64) -> Int {
     var cursor: MPMediaQuery? = nil
     var filter: MPMediaPropertyPredicate? = nil
